requirements.txt
aiohttp==3.9.5
celery[redis]==5.3.1
deepgram-sdk==2.12.0
ffmpeg-python==0.2.0
flask==3.0.3
gunicorn==22.0.0
httpx>=0.27.0,<0.28
openai>=1.40.0,<2
pydub==0.25.1
python-dotenv==1.0.1
python-magic==0.4.27
redis==5.3.0
requests==2.31.0
sentry-sdk==2.19.0
twilio==9.8.0
uvicorn==0.30.0
websockets==12.0
boto3==1.34.162
prometheus-client==0.20.0
setuptools>=68.0.0
wheel>=0.42.0
fastapi==0.115.0

# --- Added for audit fix ---
flask-sock>=0.7.0,<0.8      # used by streaming_ws.py
SpeechRecognition>=3.10,<3.11  # used by asr_service.py, duplex_voice_controller.py
phonenumbers>=8.13,<9       # used by contacts_loader.py
numpy>=1.26,<1.27           # used by realtime_voice_engine.py; conservative pin for py3.11 wheels
